+++ ProcessStore/src/main/java/at/fhjoanneum/ippr/processstore/persistence/entities/ProcessStoreBuilder.java
@@ -19,17 +19,21 @@ public class ProcessStoreBuilder implements Builder<ProcessStoreObject> {
     private Double processPrice;
     private String processApprover;
     private String processApproverComment;
-    private boolean isApproved = false;
+    private Boolean processApproved = false;
     private Date processApprovedDate;
+    private byte[] processFile;
 
-    public ProcessStoreBuilder(String processName, String processDescription, String processCreator, Date processCreatedAt, Long processVersion, Double processPrice, String processApprover, String processApproverComment, boolean isApproved, Date processApprovedDate) {
+    public ProcessStoreBuilder(String processName, String processDescription, String processCreator, Date processCreatedAt,
+                               Long processVersion, Double processPrice, String processApprover, String processApproverComment,
+                               Boolean processApproved, Date processApprovedDate, byte[] processFile) {
         checkArgument(StringUtils.isNotBlank(processName));
         checkArgument(StringUtils.isNotBlank(processDescription));
         checkArgument(StringUtils.isNotBlank(processCreator));
         checkNotNull(processCreatedAt);
         checkNotNull(processVersion);
         checkNotNull(processPrice);
-        checkNotNull(isApproved);
+        checkNotNull(processApproved);
+        checkNotNull(processFile);
         this.processName = processName;
         this.processDescription = processDescription;
         this.processCreator = processCreator;
@@ -38,8 +42,9 @@ public class ProcessStoreBuilder implements Builder<ProcessStoreObject> {
         this.processPrice = processPrice;
         this.processApprover = processApprover;
         this.processApproverComment = processApproverComment;
-        this.isApproved = isApproved;
+        this.processApproved = processApproved;
         this.processApprovedDate = processApprovedDate;
+        this.processFile = processFile;
     }
 
     @Override
@@ -50,9 +55,10 @@ public class ProcessStoreBuilder implements Builder<ProcessStoreObject> {
         checkNotNull(processCreatedAt);
         checkNotNull(processVersion);
         checkNotNull(processPrice);
-        checkNotNull(isApproved);
+        checkNotNull(processApproved);
+        checkNotNull(processFile);
 
         return new ProcessStoreObjectImpl(processName, processDescription, processCreator, processCreatedAt,
-                processVersion, processPrice, processApprover, processApproverComment, isApproved, processApprovedDate);
+                processVersion, processPrice, processApprover, processApproverComment, processApproved, processApprovedDate, processFile);
     }
 }