+++ ProcessStore/src/main/java/at/fhjoanneum/ippr/processstore/persistence/objects/ProcessOrgaMappingObject.java
@@ -1,0 +1,4 @@
+package at.fhjoanneum.ippr.processstore.persistence.objects;
+
+public interface ProcessOrgaMappingObject {
+}