+++ ProcessStore/src/main/java/at/fhjoanneum/ippr/processstore/services/ProcessStoreServiceImpl.java
@@ -2,16 +2,25 @@ package at.fhjoanneum.ippr.processstore.services;
 
 import at.fhjoanneum.ippr.commons.dto.processstore.ProcessStoreDTO;
 import at.fhjoanneum.ippr.processstore.persistence.entities.ProcessStoreObjectImpl;
+import at.fhjoanneum.ippr.processstore.persistence.objects.ProcessStoreObject;
 import at.fhjoanneum.ippr.processstore.repositories.ProcessStore;
 import com.google.common.collect.Lists;
+import org.apache.jena.base.Sys;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.ByteArrayResource;
+import org.springframework.core.io.Resource;
+import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.scheduling.annotation.AsyncResult;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import javax.transaction.Transactional;
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+import java.nio.file.Files;
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
@@ -92,12 +101,18 @@ public class ProcessStoreServiceImpl implements ProcessStoreService {
 
         final List<ProcessStoreDTO> processes = createProcessStoreDTOList(results);
 
-        return new AsyncResult<List<ProcessStoreDTO>>(processes);
+        return new AsyncResult<>(processes);
     }
 
     @Override
-    public Future<List<ProcessStoreDTO>> findAllProcessesByOrganisationId(String organisationId) {
-        return null;
+    public Future<List<ProcessStoreDTO>> findAllProcessesByOrgaId(String orgaId) {
+
+        final List<ProcessStoreObjectImpl> results = processStore.findAllProcessesByOrgaId(orgaId);
+
+        final List<ProcessStoreDTO> processes = createProcessStoreDTOList(results);
+
+        return new AsyncResult<>(processes);
+
     }
 
     @Override
@@ -139,5 +154,31 @@ public class ProcessStoreServiceImpl implements ProcessStoreService {
             processes.add(dto);
         }
         return processes;
+    }
+
+    @Override
+    public void saveProcessFile(final byte[] processFile, final Long processId) {
+        ProcessStoreObjectImpl process = processStore.findProcessById(processId);
+
+        if(process != null) {
+            process.setProcessFile(processFile);
+            //Increment version
+            Long incrementedVersion = process.getProcessVersion() + 1;
+            process.setProcessVersion(incrementedVersion);
+            processStore.save(process);
+        }
+
+    }
+
+    @Override
+    public Future<Resource> getProcessFile(final Long processId) {
+        ProcessStoreObjectImpl process = processStore.findProcessById(processId);
+
+        if (process != null) {
+            ByteArrayResource fileResource = new ByteArrayResource(process.getProcessFile());
+            return new AsyncResult<>(fileResource);
+        } else {
+            return null;
+        }
     }
 }